requirements_benchmark.txt
# PowerSGDベンチマーク機能に必要な依存関係

# 基本ライブラリ
numpy>=1.21.0
pandas>=1.3.0
pytest>=7.0.0

# 可視化・分析
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
kaleido>=0.2.1

# システム監視
psutil>=5.8.0

# 設定ファイル
PyYAML>=5.4.0

# 実験管理（オプション）
mlflow>=1.20.0

# データセット
Pillow>=8.3.0
